refactor: 글로벌하게 쓰이는 상수는 LottoConstant로 따로 빼서 관리
