+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,0 +1,8 @@
+package lotto.constant;
+
+public class LottoConstant {
+    public final static int LOTTO_PRICE = 1000;
+    public final static int LOTTO_NUMBER_LENGTH = 6;
+    public final static int LOTTO_NUMBER_MIN_VALUE = 1;
+    public final static int LOTTO_NUMBER_MAX_VALUE = 45;
+}
