+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,16 +1,16 @@
 package lotto.domain;
 
+import static lotto.constant.LottoConstant.LOTTO_NUMBER_LENGTH;
+import static lotto.constant.LottoConstant.LOTTO_NUMBER_MAX_VALUE;
+import static lotto.constant.LottoConstant.LOTTO_NUMBER_MIN_VALUE;
+import static lotto.constant.LottoConstant.LOTTO_PRICE;
+
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class LottoGenerator {
-
-    public final static int LOTTO_PRICE = 1000;
-    public final static int LOTTO_NUMBER_LENGTH = 6;
-    public final static int LOTTO_NUMBER_MIN_VALUE = 1;
-    public final static int LOTTO_NUMBER_MAX_VALUE = 45;
 
     public List<Lotto> createLottoBundle(int money) {
         int numberOfLottoToCreate = money / LOTTO_PRICE;
